r1821: * make sure new la_size is written to meta data early  * test on feature instead of version code (TASK_ZOMBIE)  * add "verify_ips" flag to drbdadm commands, so it annoys us with warnings only part of the time.  * don't report the "is ip present" bash magic as failed command  * remove .tmp_versions on clean  * minor cleanups and adjustments

+++ user/drbdsetup.c
@@ -50,7 +50,7 @@
 
 #include "drbd_limits.h"
 
-#define ARRY_SIZE(A) (sizeof(A)/sizeof(A[0]))
+#define ARRY_SIZE(A) (int)(sizeof(A)/sizeof(A[0]))
 
 /* Default values */
 #define DEF_NET_TIMEOUT             60      //  6 seconds
